Update patch set 4
Patch Set 4:

(6 comments)

Thanks for the reviews!

I would prefer to support native prometheus as it's simpler and less error prone with statsd_exporter. I can work on an abstraction and propose a 'monitoring' driver that could be implemented by statsd or prometheus_client.

@Tobias, about metrics mapping, what do you think of this apis:

  class MonitoringDriver

    @abstract
    def metric(type, value, name, template, help, **kwargs):
      ...

    def gauge(name, value, template, help, **kwargs):
      self.metric('gauge', value, name, template, help, kwargs)


  class Statsd(MonitoringDriver):
    def metric(type, name, value, template, help, kwargs):
      self.statsd.__getattr_(type)(template.format(**kwargs), value)

  class Prometheus(MonitoringDriver):
    def metric(type, name, template, help, kwargs):
      gauge = prometheus_client.Gauge(name, help)
      gauge.labels(**kwargs).set(value)


Would such things be accepted?

Patch-set: 4