Change the remaining time fallback in JS
The Python set of calculating the remaining time deals with the fact
that if not jobs are queued, the remaining time is None. The JS seems to
set the fallback to true, which the time JS function deals with by
casting to int, and setting it to 0. Change the fallback to 'undefined'
which firstly, is true, and secondly, the function can deal with.

Change-Id: I3e1ea00084a0dc9497b9678b816883d9b984312b