Update patch set 6
Patch Set 6: Code-Review-2

This doesn't match the data model.  Different variants of different jobs can have different parents, but the job list returns only a single item for each job (which is the right thing to do).  The parent-child relationship is at the job variant level, and that is provided by the /job/ api endpoint.

Patch-set: 6
Label: Code-Review=-2