wip: Assume config change if file list is incomplete
The Github pull requests files API only returns at max the first 300
changed files of a PR in alphabetical order.

In case the PR has more than 300 changed files, it might be that Zuul
is not considering changes to the config and not using them
speculatively due to the incomplete file list.

Change-Id: I2efb4f5d06a0815167aa590d24dc8277bb278ae8