Update patch set 4
Patch Set 4:

I think we should do both. There is already an upstream PR [1] that tries to fix it but it seems to be a bit controversial and stuck so we should at least have an intermediate solution. After the upstream fix we can just throw away the callback module and the special handling in the executor and still have the unreachable tests (which didn't exist so far).

[1] https://github.com/ansible/ansible/pull/36660

Patch-set: 4