wip: Use merger to get list of files for pull-request
The Github pull requests files API only returns at max the first 300
changed files of a PR in alphabetical order.

In case the PR has more than 300 changed files, it might be that Zuul
is not considering changes to the config and not using them
speculatively due to the incomplete file list.

Also jobs using file filters might not be considered in case the
relevant files are not in the list of files for a change.

Change-Id: I10a593e26ac85b8c12ca9c82051cad809382f50a