Update patch set 1
Patch Set 1: Code-Review-1

(4 comments)

Very cool idea, and I could definitely use this. I'd prefer that Zuul owns its own auth. Offload to the owning webserver is an ok idea, but it causes a lot of unnecessary complexity in practice. In fact what I'd really like to see is a spec that discusses the auth problem entirely on its own, with this as the first consumer of the auth. JWT, Oauth2, SAML, are all valid choices here, and would be far simpler implemented directly in zuul-web.

Patch-set: 1
Label: Code-Review=-1