Add a zuul-local tool for testing playbooks
This is a horrible tool for doing horrible things, but it has helped me
do at least half of the very difficult job of trying to simulate the
zuul-executor while trying to iterate on complicated playbooks.

We should not merge this. This is just an example. Instead we should
make a proper entry point that makes the cli tool fetch the zuul
configuration from zuul's API (also non-existant as of now) and uses
that to run the playbooks, requiring users to provide site variables and
secrets, and substituting in local values for executor details,
pipelines, source paths, etc. Ideally this would also run the playbook
in bubblewrap for the full experience.

But, I wanted to get this up here in case anyone else is struggling.

Change-Id: I7ed1f0afcdb9383feda3086f94d9c77bbdfac8c7