WIP - Pagure driver - https://pagure.io/pagure/
This change adds a Pagure driver for Zuul. The driver is for
Pagure https://pagure.io/pagure/. Pagure is a Github like forge.

This driver is usable with Pagure >= 5.3.

For history: first Pagure PR gated by
Zuul : https://pagure.io/test-zuul/pull-request/22

The driver is usable with project on pagure.io.

It is still a WIP a the patch lacks unittests and doc.

Change-Id: I1711653355ae26a3fff3bb6de3c6fca7113cdd01