Update patch set 3
Patch Set 3:

It might be nice to be able to link to this from the status page. The change links currently go to the source of the change - which I think is good. Is there a way that doesn't blow to add a link to the change status page for a change from the main status page?

Patch-set: 3
Label: Verified=0