WIP: Add API endpoint to get frozen jobs
This allows a developer to query what the jobs will look like for a
project against a pipeline for a branch tip. It will return the list of
roles and playbooks that will be used in the job execution.

The idea is that this will be consumable as a way of reconstructing jobs
locally for debugging.

Change-Id: I89afac1fd4ebe7924ad60dfc689dc9018726c133