Reorganize admin manual
This makes a number of changes to the installation/configuration
documentation in the admin manual.

Remove quick-start guide.  The process of quick-starting is
covered by the installation and setup tutorial, which is now the
first of the installation sections.  The reference material from
quick-start is now in the tutorial.

Rename the tutorial quick-start.  It's nice to have something
named quick-start, and the tutorial fits the bill.

Rename the installation section "Installation Reference".  This
now has more detailed information about installation and
deployment choices, but has very little procedural documentation.

Make zuul-from-scratch more internally consistent in style (use
code-block:: shell and heredocs wherever possible).

Change-Id: I7e4714ce5e775dc9ac0988c3470eef1f74fb36d6