Add a quick-start test job
This playbook performs approximately the same steps documented in
the quick-start tutorial.  Use it to verify that things still function.

Note that it does not yet build the images from the current source
(rather, it downloads them from dockerhub) so it is not suitable
as a gate check (it does not use the change under test in zuul).
It does, however,  use the sample config files from the change under
test, so can be used to verify changes to those.

Incidentally, this is the first-ever live functional test between
Zuul, Nodepool, and Gerrit.

Change-Id: I5b3dc4b8a8d409787d07b4ad155898f97f1e9eb9