Make update interval configurable at compile time
Currently the update interval of zuul-web is fixed to 5s and tailored
to the big deployment of openstack. In many smaller or multi-tenant
deployments this can be easily a lower value so the user has the
feeling of faster reaction times of zuul. This makes this timeout
configurable at compile time so a deployer of zuul can change this
without patching zuul.

Change-Id: I0fa71c3b8fd8f697a713d1e5fff0f87b20ba45cd