Update patch set 2
Abandoned

We patched this months ago in our deployment because of a user request. Now our deployment also grew so I will check with my collegues if we still really need that. (I personally doubt it, but in a small fast zuul it felt just much snappier). So abandoning for now.

Patch-set: 2
Status: abandoned