Update patch set 1
Patch Set 1:

If I'm following correctly, this will give the version of the zuul-web process.  Currently the status page (and api endpoint) returns the version of the scheduler.  If we only show one version, I think it should be the scheduler (since it's the SPOF that is hard to restart).  If we show more than one version, we should probably show them all.  I was thinking that would be easiest when all the components connect to zookeeper (we can just query the zk registration for each component).  But if we wanted to do it earlier, we could probably do something with Gearman, though I think that would get complicated.

If you just want to do scheduler and web for now, we should probably have the info endpoint supply both.

Patch-set: 1