Use zuul/web/static in setup_hook
The actual location of the files we want to put into the wheel are the
files at zuul/web static. web/build is just a symlink so that the
javascript build system in web/ puts files where we want them without
needing to do build config things.

There is no actual difference - python processes the symlinks just as
well as it processes the direct paths. But from a semantic perspective,
pointing to the paths as they will be in the wheel makes a little more
sense than pointing to paths through a convenience symlink, when trying
to sort out what's going on and why.

Also, add debug statements to verify the __init__.py files we're
creating in log output.

Change-Id: Ib93b0eb22f0096b33bcfdb862a47765e48be7203