Update patch set 1
Patch Set 1: Code-Review-1

(8 comments)

I like that generally.

Is this meant for the general zuul user or the zuul-of-openstack user? I assume it's the first otherwise this should be in some openstack doc repo.

What I'm struggling with is that while I like this howto it leaves out the the information about the used base jobs. It howto strongly relies on specific functionality that must be provided by the base job. And I assume that (especially for the elasticsearch) you target for the openstack or software factory base job?

Patch-set: 1
Reviewer: Gerrit User 16068 <16068@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1