Update patch set 1
Patch Set 1: Code-Review-2

As a technology test, this is a success for me -- it renders in my browser and is relatively responsive.

However, as information display, this method of visualization doesn't convey any information not already in the current status page, and, in fact, conveys much less and in a confusing manner.  There may be a place for this kind of display, but I don't think it's going to be the status page.  The most likely place to explore is the project-pipeline configuration, where showing a graph relationship between jobs on an individual project may be scalable and useful.

Patch-set: 1
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-2