Update patch set 1
Patch Set 1: Code-Review-1

We *already* report the scheduler version on the status page.  If you add this, we will report two different versions in two different places with no explanation for the difference.  If we report only one version, it should be the scheduler (as it is today) because that's what determines what job features are available to users.  If we're going to add more components, such as web in this change, we should put all the version info into one place and identify which components the version corresponds to.

So if you're going to do this, the minimum necessary is to also include the scheduler version in the same place.

Patch-set: 1
Label: Code-Review=-1