Add reference to svg icon in the PWA manifest
With progressive web apps you can specify one or more icons to be used
to identify the web app in scenarios such as saving the app to a phone
home screen.

  https://www.w3.org/TR/appmanifest/#icons-member

Add a link to the svg version so that we get a zuul logo when adding
zuul to the home screen of a phone.

Change-Id: I0ea47aa341457c4816154b0272f6e460f3faad74