Add icons in the PWA manifest
With progressive web apps you can specify one or more icons to be used
to identify the web app in scenarios such as saving the app to a phone
home screen.

  https://www.w3.org/TR/appmanifest/#icons-member

Additionally, the google lighthouse recommendations:

  https://developers.google.com/web/tools/lighthouse/audits/install-prompt#recommendations

Indicate "icons must include a 192px and a 512px sized icons"

Add 192px and 512px png logos as well as a link to the svg version so
that we get a zuul logo when adding zuul to the home screen of a phone.

Change-Id: I0ea47aa341457c4816154b0272f6e460f3faad74