Update patch set 2
Patch Set 2: Code-Review-1

Not crazy about the log rework - setup_logging needs to be called before self.log - so extracting it to that property could lead to using it before setup and not noticing (it'll throw an undefined error currently) To do a property like that, it would want to include setup_logging, and probably want to save the result into a self._log wrapped in an if not self._log: ... but then we'd need to define a default for self._log - and that's then a lot more code.

Let's talk about the extract defaults direction before you go any further with it.

This feels like a code rearrangement patch - but the direction you're looking at going with a defaults module might be nice.

It's helpful in commit messages to start with the why or the problem the patch is trying to solve - that helps us as reviewers with context.

For instance, if there is an end-state here where by extracting defaults we could sync them with the config documentation to reduce skew, or if we need to extract them to make something easier for testing, or similar, those are potentially good goals and worth discussing.

Probably a good idea to chat with us about it in #zuul or on zuul-discuss@lists.zuul-ci.org

Patch-set: 2
Reviewer: Gerrit User 2 <2@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1