Update patch set 2
Patch Set 2:

> Hi! Thanks for the patch.
 > 
 > The Zuul project typically doesn't accept patches that are just
 > codestyle changes. They make tracing coding intent through git
 > blame later harder, as well as creating merge conflicts for other
 > outstanding patches. Minor codestyle fixes are fine when they're
 > part of a functional patch that touches the same code.

I agree, that this patch is not something functional or important, 
but while I learn codebase, I often prepare some relatively light
and safe patches to raise some question and clarify different
aspects of application.

In this case I have some question about code style which looked 
strange for general python code guidelines. Thank you for 
clarification on why this done in this way and why this is not 
changed yet.

 > In this case specifically, the intended prevailing style in the
 > codebase is camelCase for methods and snake_case for properties. 

Actually, no. From what I've already seen here, it's hard to state that 
camelCase is for methods and snake_case is for properties. 
A lot of methods have snake_case too.
That's why I thought that it's just a kind of legacy stuff from "old days".

 > I don't think it's worth going through and changing the places where
 > things diverge from that style.

Ok. Agree.

But zuul project uses very uncommon code style which is not 
documented anywhere. I've seen some notes in [flake8] section of 
tox.ini, but they not include rules which you mentioned in your review.

Patch-set: 2