Merger: automatically add new hosts to the known_hosts file
To improve the experience for new users, have the mergers automatically
add new ssh hosts to their known hosts files.  They will still validate
existing entries, so a paranoid sysadmin can still pre-populate the file
to avoid trojan attacks on the initial connection.  New users in the
typical case of a secure internal network will not need to manually add
host keys to start Zuul.

This matches the behavior of the scheduler with its stream-events
connection to Gerrit.

Change-Id: Ib416bbb2ffcfedd5d114709af3931e154a1a92ac