Fix issue in Github connection with large diffs
When a PR consist of a large number of changes github3.py will raise an
exception, since Github will fail to generate the diff.

This case can be handled via the same mechanism that is used for getting
the list of files when Github doesn't return the full file list (e.g. if
there are more than 300 changed files in a PR)

Change-Id: Ic1dd6d381536268036dd845ecb5e2073383704a0