Update patch set 3
Patch Set 3:

For the record -- this is not a behavior difference between gerrit and github, this is the difference between triggering on the change/pr merged event versus the ref-updated event.  In both drivers, change/pr merged has the change/pr as the ref, while the ref-updated event has the branch as the ref.  Supercedent was written for the ref-updated use-case.

Patch-set: 3