Add reenqueue utility
Add a simply utility script which turns an inventory file from a
completed build for a ref object into the command needed to
reenqueue that ref into the same pipeline. This is convenient for
when tag-based release jobs or post-merge docs jobs fail due to
outside influences and you want them rerun.

Change-Id: Id38e7ba62c4a3698fb503584dbed0379085736df
Signed-off-by: Paul Belanger <pabelanger@redhat.com>