Update patch set 3
Patch Set 3: Code-Review+2

Is it still necessary to set the end-time in the reporter?

I'm trying to think of a case where the reporter may be used mid-job, ie before it has actually 'ended'. I don't think this is something we'd see currently, but it may be something that could happen in the future.

For now though, this change looks okay to me, but if it isn't necessary to set in the reporter it'd probably be better not to.

Patch-set: 3
Reviewer: Gerrit User 7069 <7069@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2