Display ref instead of NA for time triggered items
Currently time triggered items show up as 'NA' in the status ui. If we
want to dequeue such an item we need the ref which is not existing
anywere in the status, not even in the pure json. So add the ref if
it's there and display the ref instead of 'NA' in the status ui if the
ref is set.

Change-Id: Iedae91d96ab91f4b4131da66ee317072f38e376b