Update patch set 2
Patch Set 2:

So, I'm going to push back on this one. This code is well covered in the current tests, so it won't cause any explosions any time soon. But testing that this happens is an exercise in futility IMO. Either we have to develop a whole new way to run playbooks in a container in the unit tests so we can move python somewhere else, or we have to test this outside the unit tests. I'm open to suggestions, but frankly, having this is more important than having it 100% covered by explicit test coverage.

Patch-set: 2