Add a statsd check for clashing keys
This added unit-test check ensures that when we assert a statsd key is
set, we haven't tried to set any other subkeys under existing counter
or gauge keys.  Although this is not 100% reliable as different tests
could still create conflicting keys, this would have at least found
the problems in I127e8b6d08ab86e0f24018fd4b33c626682c76c7 and thus
rules out one sort of problem.

Fix the existing statsd test-case which is trying to set subkeys that
wouldn't work; rename them as we're we're testing the hostname
expansion.

Also add a negative test to ensure the failure is found when
asserting.

Change-Id: Ie1ca39659d32e45e0ad1d80ca55e28ae6568a8b4