web: refactor jobs page to use a reducer
Thios change updates the status page component to dispatch reducer
action instead of direct axios call. This enables using the generic
error reducers as well as keeping the jobs list in the store to
avoid repeated query. This change also refactor the Job list into
a container and extends the common refresh component.

Change-Id: I7f44e83d6bea8ee915c81d7ba9afd6d9a4c89d38