Update patch set 1
Patch Set 1:

One thing we're missing which will need to be formalized is the actual list of maintainers. I'm assuming it's roughly the set of names at https://review.openstack.org/#/admin/groups/zuul-core but either within this change or a follow-up we need an official record of who they are (and ideally also how to reach them for purposes of direct communication when needed).

Patch-set: 1
Label: Verified=0