Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

> 1) What veto rights do the leader and/or maintainers have? Both in respect to code and also governance policies.

Personally I'd hope this comes under the "collaborate and consider other than yourself".  If a change was proposed that fundamentally broke your usage of Zuul, such to the point that you feel you need to veto it,  it's hard to imagine it would be getting votes to merge.  Similarly if you're pointing out "this is a huge security hole" or something, it doesn't seem like other maintainers would want to merge that.

-1 from me because I don't see in here any discussion of the specs process (I really hope I didn't miss it); which is already active for things like zookeeper/etcd, just to formalise what's expected there.  As mentioned inline, I think it could cover Josh's veto points too.

Patch-set: 1
Label: Code-Review=-1