Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

Note that starting build times are normally expected to be *highly* variable, depending on the job.  It's almost like trying to determine performance by looking at the times for all builds.  However, I think in very large aggregate over a long period of time, you might be able to glean information from this.

However, if I saw this graph on OpenStack, and saw that it got higher over the past 6 months, I would be very include to suspect that the bulk of the increase was just due to more people writing larger native v3 jobs with more required-projects.

I'm not opposed to this, but I'm not excited about it because of the problems interpreting it.

The -1 is for the error inline.  Might be good to add it to a test -- it would have caught that.  :)

Patch-set: 2