Update patch set 1
Patch Set 1:

(1 comment)

Thanks Jim for the draft -- I like the simplicity of this and how well it reflects how Zuul has been operating so far.

As I mentioned before, my main concern is checking adherence to the "open community" principle, which basically makes sure that contribution is the only currency, and anyone can rise to leadership positions. The risk with a model of self-selecting maintainers is to create an aristocracy, disconnected from contributors, eventually leading to a fork of the project. This has happened before in similar setups. As Zuul gets confirmed as a top-level project under the OSF, we are asking the Foundation members to invest in establishing and marketing Zuul's brand, so avoiding forks is important.

In Zuul's proposed governance, it will depend a lot on how easily active contributors will get accepted as maintainers. Is the intent is to closely track active contribution and aggressively make people maintainers ? Or to have the smallest group possible ? How actively do you plan to cycle out inactive contributors ?

Since there is no governance body above the Zuul maintainers, ideally I'd like to see something in place to prevent any aristocratic drift. It does not mean adding complexity, boilerplate process or changing how you operated so far: it could be a safety valve system. For example, a petition of more than 50% of commit authors representing more than 66% of the changes over the past year could call for a governance reset. This safety valve should basically never trigger if maintainers are correctly... er... maintained. It should therefore not add any boilerplate process and retain the spirit of what you wrote.

Is such a safety valve something you'd be open to add ?

Patch-set: 1
Label: Workflow=0