Update patch set 2
Patch Set 2: Code-Review-1

The space should be outside of the span which should just be the number. However that would require refactoring, a trailing space on one of the proceeding lines, or a &nbsp;. I'm not sure what is desirable and I'm happy to be out voted. (I'd probably lean towards a &nbsp;).

Patch-set: 2
Label: Code-Review=-1