Update patch set 6
Patch Set 6: Code-Review+2

This is really cool :-)

The zuul-build-dashboard job doesn't point to the right backend though. The API page just calls the hardcoded endpoint. This could be fixed later though or might be difficult to avoid.

I wonder if we want to store the swagger.json elsewhere to make it more prominent part of developing the API. Additionally if we ever want to make use of any of the swagger libraries (to help with routing for example) we might need it elsewhere. Probably something we can visit later though if needed.

Patch-set: 6
Label: Code-Review=+2