Update patch set 1
Patch Set 1:

(2 comments)

Joshua, I don't think the project lead has any special veto rights;
the maintainers do have veto rights with -2 though.  It's mentioned
under "approve or reject" but I'm happy to clarify that.

I also agree that we should codify the additional maintainers of
zuul-jobs.

Ian, I would prefer to avoid mention of specs here.  I'm trying to
keep this document as minimal as possible.  Specs are one of the ways
we collaborate and achieve consensus, but they are one among many, and
they are, after all, a form of code review which is covered.  The main
thing that I think is important in this document is that we codify
that the process is collaboration and consensus.

Thierry, I understand where you're coming from -- I've seen the same
kinds of project splits over the years.  Fundamentally we're talking
about the case of a complete breakdown in collaboration.  Whether the
result of that is forcible expulsion of long-term project
participants, or a fork created by newer contributors, the results are
equally catastrophic.  I don't want either to happen and I would
prefer to avoid codifying the nuclear option.

I do think we need to guard against the divergence you describe, so I
suggest we do the following: add mention that recruiting new
maintainers is desirable (that we should not have artificial limits on
the number of maintainers) and indicate that the project lead is
available to guide contributors into maintainership positions.  I'm
not sure if we need to codify this or not, but the current practice
seems to be a significant grace period in the removal process, as
we're trying to accomodate peoples changing circumstances and habits.

Patch-set: 1
Label: Verified=0