Update patch set 2
Patch Set 2:

> Thierry, I understand where you're coming from -- I've seen the
 > same kinds of project splits over the years.  Fundamentally we're
 > talking about the case of a complete breakdown in collaboration.
 > Whether the result of that is forcible expulsion of long-term project
 > participants, or a fork created by newer contributors, the results
 > are equally catastrophic.  I don't want either to happen and I would
 > prefer to avoid codifying the nuclear option.

I agree that the results are equally catastrophic from the perspective of the maintainers team. The safety valve would really only serve to protect the investment of the Foundation members, by guaranteeing that the brand of the project stays with the largest group of contributors. By becoming a top-level project of the OSF, "Zuul" will receive some long-term investment from the Foundation members through their investment in the OSF -- the safety valve would guarantee that that investment is not wasted if things do not go as planned. With the currently-proposed governance, you are asking those members to invest long-term in the maintainers team (current and future), rather than in Zuul-the-project... a subtle distinction which they may or may not object to. Adding a carefully-calibrated safety valve would IMHO preserve the intent of this proposal while avoiding this potential obstacle.

Patch-set: 2