Update patch set 15
Patch Set 15: Code-Review-1

(1 comment)

This is looking great! That docstring in the pagureconnection.py is fantastic - and seems like it should be documentation.

Left a -1 for the copyright header. This is exciting work to see.

Patch-set: 15
Reviewer: Gerrit User 2 <2@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1