Update patch set 16
Patch Set 16:

Thanks for the initial review. I'll move that doc in the real Zuul doc in a followup PS. Yes that's exciting to integrate Zuul and Pagure. It requires some changes on Pagure but the Pagure team is always happy to help and reactive when features/changes are proposed. I have good feeling that we'll get soon on fully usable driver :)

Patch-set: 16