Interpret ANSI while streaming logs
Many programs think that the console is still a terminal and spew out
ANSI color sequences. We might as well display those in a nice way while
streaming the logs.

Note that adding the numerous spans that color changes require via the
ref method was fairly complicated. As a result, we refactor the code
here to work using React's built in state tracking, drastically
simplifying the code, with no noticable impact on browser CPU or RAM.

Change-Id: I39787e7ef1f25447c881980e3e56652a690625cc