Cache empty branch config to prevent spurious reconfig
There is an edge case in github handling where it's possible to
change a branch from unprotected to protected without Zuul being
notified, therefore if we see an event on a branch which has no
configuration cached, we perform a tenant reconfiguration in order
to catch this case.

We also have an optimization where if we load no configuration from
a repo whatsoever, we don't cache anything for it.

Therefore, every time we see an event on such a repo, we perform
an extra configuration (because the repo has no cached data).

To avoid this, explicitly cache an empty configuration object on
every project-branch we know about.  This way only a branch which
has gone from unprotected to protected will show up with no cached
data and trip the reconfiguration.

Change-Id: I8161b571b03c4d8222abd6f1503190917331f422