Update patch set 6
Patch Set 6: Code-Review-1

I know it's long, but it might still be good to run the test against the whole output just to make sure we don't add unexpected things in the future. It'll also keep us aware of the size concerns that James has pointed out. But if you disagree this otherwise lgtm.

Patch-set: 6
Label: Code-Review=-1