Update patch set 7
Patch Set 7:

I like this. However there seems to be no coupling to the code or validation in test cases. So I fear that this documentation will become outdated pretty quickly especially because every change in any route will need the appropriate change in the swagger.json which is almost guaranteed to be forgotten in the future.

Patch-set: 7