Update patch set 2
Patch Set 2:

> Shouldn't the document explains the Open Design of Zuul? I
 > personally find it unclear how requirements are gathered, what is
 > the roadmap, when is the next release etc.. For example the k8s
 > nodepool driver proposal stalled without comments for more than 10
 > months.

This document is meant to codify a set of existing practices to form an initial seed policy by which we will be bound as a community. It does state that design discussions are held and decisions made in public, and that anyone is welcome to participate. I don't think it's the place to describe detailed process such as requirements gathering, roadmaps and release process. The barrier to modification for this document is intentionally quite high, and so limiting what we include allows for more fluid definition of our operating processes independent from (but guided by) this work of policy.

Patch-set: 2