Update patch set 4
Patch Set 4: Code-Review-1

(1 comment)

Please see the inline comment. We also need to test this.

In order to do this you could fork the test case test_unprotected_branches and enhance the FakeRepository in fakegithub.py such that it can return 404 in get_url_branch_list if protected is enabled.

Patch-set: 4
Label: Code-Review=-1