Use urljoin for git driver getGitUrl
The current code uses a format string with an unconditional injection
of a /. This can result in a double-slash which can then cause git to
be unable to clone things. That in turn can lead to an indefinite hang
of a job trying to use the repo from the git driver.

Replace the format string with urllib.parse.urljoin which does the
right thing with trailing slashes in baseurl in either direction.

While we're in there, switch lsRemote from os.path.join to
urllib.parse.urljoin just on principle, since it is ultimately a url.

Change-Id: Id0e529b916ff75d6b6df98d978d2349f6037499c