Update patch set 3
Patch Set 3:

Turns out this change is not necessary. Most of the AnsibleJob code is needed for the zuul-runner execute command. To reduce the amount of moving code, I updated the parent change to only move what is re-usable out of the executor module to the common module. This change is no longer needed.

Patch-set: 3