Return the dashboard build url for builds with no URL
When the build doesn't have a URL, we currently return finger://
urls, which exist for commuication with the log streamer. Ther're
never a valid thing to return, as by the time they are returned,
they've ceased to be valid.

Job should be returning their log location to be returned to the user.
But sometimes, like when POST_FAILURE or DISK_FULL occurs at the wrong
time, the user can be left with a finger:// url that isn't useful.

While it's not yet time to just always return a build url from the
dashboard, the dashboard build url is at least maybe a better
experiene than a broken invalid finger url.

Add config settings to the [web] section for setting the web base
url (for the simple case) and a build_url_format string for the
more complex case like a whilelabel setup where
{web.url}/t/:tenant/build/:build is not the correct location.

Change-Id: Ibfe5fa8946e3863a54639f966059849629fd62a6