Update patch set 9
Patch Set 9:

The new release of sphinxcontrib-openapi does not include the example output, though the previous build with the git sha did.  What's the story there?

If I understand correctly, we will expect developers to edit the swagger file manually.  Assuming that's the case, I've proposed a followup to switch the file to YAML (which is a supported openapi document type) to make that easier (I don't think we should ever ask a human to edit JSON).

There are some errors in the descriptions for some of the fields, but given the proposed switch to yaml, let's also save that for a followup.

Patch-set: 9
Label: Verified=0