List changed files for all commits between refs
Using git merge-bases will not work properly when the target branch
of a PR was merged back into the PR branch. Unfortunately this seems
to be a common workflow for some peope to keep a pull request up to
date with e.g. master.

Using git rev-list will return the set of commits that are reachable
from the PR head but without those already in the base branch.

Change-Id: I25c75a70892bbac52db783fb63513f6aa37cb5f6