Build zuul containers with dockerfile not pbrx
While pbrx is nice and all, it's quite the divergence from how
the rest of the container ecosystem works. Switch to using
Dockerfile and the python-builder image.

Use install-js-tools.sh in Dockerfile

Add curl and gnupg2 to install-js-tools.sh because it doesn't
work without them and they're not in python:slim by default.

Depends-On: https://review.openstack.org/631878
Change-Id: I84f6dd2a8e3222f7807103dcbb61bdadedfdd22d