WIP Build zuul containers with dockerfile not pbrx
While pbrx is nice and all, it's quite the divergence from how
the rest of the container ecosystem works. Switch to using
Dockerfile and the python-builder image.

We need to update this with build jobs, as well as with some
more specific dockerfiles for individual components. This is
mostly to show the simple version of using the python-builder image.

https://review.openstack.org/631878
Change-Id: I84f6dd2a8e3222f7807103dcbb61bdadedfdd22d