Stop building an explicit zuul-base image
We have a utility image that we use for running the zuul command
that doens't have any additional software installed. Although it does
set a COMMAND of /usr/local/bin/zuul, it could still be useful as
a general base image for other people if they wanted such a thing.

Change-Id: I4f92d8707f478c73cd41678db4ed511e8dbfbb7e