Update patch set 2
Patch Set 2: Code-Review-1

I have two concerns with this:

First, I'm concerned about keeping parity with other reporters.  Certainly we should not report to Gerrit or GitHub.  But perhaps we should report to the SQL database?  And then make it available in the web UI?  If we decide that's the case, we don't have to update this change, but we should at least have another change ready to merge which implements that.

Second, is this necessary?  Given that we probably haven't run the post playbooks, there won't be any logs to see, so I'm not sure what use it would be to an end user.  Would a simple count of the number of retries be enough to indicate that the build had problems and further investigation is necessary?  Or, do you want to have post-playbooks run to attempt to salvage logs?

Patch-set: 2
Label: Code-Review=-1